Add digital averaging

+++ src/actions/fileActions.js
@@ -144,7 +144,7 @@ export const load = () => async dispatch => {
     if (pos < buffer.length) {
         len = buffer.slice(pos, pos + 4).readUInt32LE();
         pos += 4;
-        options.bits = new Uint8Array(buffer.slice(pos, pos + len));
+        options.bits = new Uint16Array(buffer.slice(pos, pos + len));
     } else {
         options.bits = null;
     }
